backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt<4.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic[email]>=2.7.4
python-dotenv>=1.0.1
Pillow>=10.3.0
#tensorflow>=2.16.1
numpy>=2.0.0
alembic>=1.13.1
